Move instruction to a directory
